First steps in coding - more exercise's home work

+++ 01.CSharp-Basics/02.FirstStepsInCodingExercise/FishTank/StartUp.cs
@@ -9,7 +9,7 @@ namespace FishTank
             int length = int.Parse(Console.ReadLine() ?? throw new InvalidOperationException());
             int width = int.Parse(Console.ReadLine() ?? throw new InvalidOperationException());
             int height = int.Parse(Console.ReadLine() ?? throw new InvalidOperationException());
-            double percentage = int.Parse(Console.ReadLine() ?? throw new InvalidOperationException());
+            double percentage = double.Parse(Console.ReadLine() ?? throw new InvalidOperationException());
 
             double volume = (length * width * height) * 0.001;
             double perc = percentage * 0.01;
